+++ tksheet/row_index.py
@@ -41,7 +41,7 @@ from .functions import (
     is_contiguous,
     new_tk_event,
     num2alpha,
-    pickled_event_dict,
+    stored_event_dict,
     rounded_box_coords,
     try_binding,
 )
@@ -878,7 +878,7 @@ class RowIndex(tk.Canvas):
                         "displayed": disp_new_idxs,
                     }
                     if self.MT.undo_enabled:
-                        self.MT.undo_stack.append(pickled_event_dict(event_data))
+                        self.MT.undo_stack.append(stored_event_dict(event_data))
                     self.MT.main_table_redraw_grid_and_text(redraw_header=True, redraw_row_index=True)
                     try_binding(self.ri_extra_end_drag_drop_func, event_data, "end_move_rows")
                     self.MT.sheet_modified(event_data)
@@ -2344,7 +2344,7 @@ class RowIndex(tk.Canvas):
             self.fix_index(datarn)
             if not check_input_valid or self.input_valid_for_cell(datarn, value):
                 if self.MT.undo_enabled and undo:
-                    self.MT.undo_stack.append(pickled_event_dict(event_data))
+                    self.MT.undo_stack.append(stored_event_dict(event_data))
                 self.set_cell_data(datarn=datarn, value=value)
                 edited = True
         if edited and cell_resize and self.PAR.ops.cell_auto_resize_enabled: